Refactor relay types
