+++ core/services/relay/evm/ocr2vrf.go
@@ -10,7 +10,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/libocr/gethwrappers2/ocr2aggregator"
 	"github.com/smartcontractkit/libocr/offchainreporting2/chains/evmutil"
-	"github.com/smartcontractkit/libocr/offchainreporting2/types"
+	ocrtypes "github.com/smartcontractkit/libocr/offchainreporting2/types"
 	"github.com/smartcontractkit/sqlx"
 
 	relaytypes "github.com/smartcontractkit/chainlink-relay/pkg/types"
@@ -18,6 +18,7 @@ import (
 	"github.com/smartcontractkit/chainlink/core/chains/evm"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/services/ocr2/plugins/dkg/config"
+	types "github.com/smartcontractkit/chainlink/core/services/relay/evm/types"
 )
 
 // DKGProvider provides all components needed for a DKG plugin.
@@ -101,7 +102,7 @@ type dkgProvider struct {
 	pluginConfig        config.PluginConfig
 }
 
-func (c *dkgProvider) ContractTransmitter() types.ContractTransmitter {
+func (c *dkgProvider) ContractTransmitter() ocrtypes.ContractTransmitter {
 	return c.contractTransmitter
 }
 
@@ -110,12 +111,12 @@ type ocr2vrfProvider struct {
 	contractTransmitter *ContractTransmitter
 }
 
-func (c *ocr2vrfProvider) ContractTransmitter() types.ContractTransmitter {
+func (c *ocr2vrfProvider) ContractTransmitter() ocrtypes.ContractTransmitter {
 	return c.contractTransmitter
 }
 
 func newOCR2VRFConfigProvider(lggr logger.Logger, chain evm.Chain, rargs relaytypes.RelayArgs) (*configWatcher, error) {
-	var relayConfig RelayConfig
+	var relayConfig types.RelayConfig
 	err := json.Unmarshal(rargs.RelayConfig, &relayConfig)
 	if err != nil {
 		return nil, err
